Introduce Huddle.
Huddle is a way to build CDDL in Haskell. It is not directly a DSL for
assembling CDDL documents, but a means to use Haskell's own abstraction
facilities to build CDDL.

It does not support all of CDDL's features. For example, generics are
not supported - one can instead use function application in Haskell for
that. Likewise, there is no explicit support for generating CDDL groups
- rather, one can simply use Haskell composition to include all relevant
parts. In this way, one should perhaps see CDDL as a compilation
target for Huddle - the generated CDDL should be correct, but not
necessarily something one might want to edit by hand.

At present, no checks are performed that a Huddle document is complete -
e.g. that there are no references to undefined types.