Ensure termination of generators.
Structure is heavily recursive, so we need to ensure there's a limit to
how far we recurse.

In particular, we need to be careful when we create generic arguments,
since they can be arbitrarily complex expressions. So we introduce
`maybeRec` which tries not to go down that rabbit hole too many times.