Introduce HuddleM
This is an attempt at addressing #35.

We provide a monad for specifying Huddle definitions, which will collect
them in definition order. This allows a fairly convenient way to put
together a specification, at the cost of it being trickier to re-use
definitions elsewhere, since they need to be returned from the monad.