Support constraining reference types
Before this commit, it wasn't possible to apply constraints (e.g.
'.size') to references, only directly to Value types. This gives us some
nice type safety (you can't apply '.cbor' to an integer, for example),
it also doesn't allow us to constrain types which are passed as generic
parameters, or where we genuinely want to constrain another rule.

This commit changes that by allowing references (both direct and
generic) to be constrained. Obviously we don't have any guarantees here
that what we are constraining makes sense, but we never had that in CDDL
anyway. And we retain the type safety for anything that _isn't_ a
reference.

This resolves issue #31